+++ packages/terra-application-name/tests/wdio/application-header-name-spec.js
@@ -1,13 +1,12 @@
 Terra.describeViewports('ApplicationHeaderName', ['tiny', 'medium'], () => {
-  describe('Displays a default application header name', () => {
-    before(() => browser.url('/#/raw/tests/terra-application-name/application-name/application-header-name-default'));
+  it('Displays a default application header name', () => {
+    browser.url('/raw/tests/terra-application-name/application-name/application-header-name-default');
     browser.waitForVisible('[class*="demo-image-container"]');
-    Terra.it.validatesElement({ selector: '#default' });
+    Terra.validates.element({ selector: '#default' });
   });
 
-  describe('Displays a truncated application header name', () => {
-    before(() => browser.url('/#/raw/tests/terra-application-name/application-name/application-header-name-truncated'));
-
-    Terra.it.validatesElement({ selector: '#truncated' });
+  it('Displays a truncated application header name', () => {
+    browser.url('/raw/tests/terra-application-name/application-name/application-header-name-truncated');
+    Terra.validates.element('truncated', { selector: '#truncated' });
   });
 });