+++ packages/terra-application-name/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
+
 ## 3.33.0 - (August 18, 2020)
 
 * Added